BiMCF/BiMCF/requirements.txt
accelerate==0.29.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
datasets==2.20.0
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
filelock @ file:///croot/filelock_1700591183607/work
flash-attn==2.6.3
fonttools==4.55.1
frozenlist==1.5.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
hf_transfer==0.1.8
huggingface-hub==0.26.3
idna @ file:///croot/idna_1714398848350/work
imgcat==0.6.0
inquirerpy==0.3.4
ipdb==0.13.13
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.3
matplotlib-inline==0.1.7
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.1.0
multiprocess==0.70.16
networkx @ file:///croot/networkx_1717597493534/work
numpy @ file:///croot/numpy_and_numpy_base_1695830428084/work/dist/numpy-1.26.0-cp310-cp310-linux_x86_64.whl#sha256=fc2732718bc9e06a7b702492cb4f5afffe9671083930452d894377bf563464a3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.11.1
pexpect==4.9.0
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyarrow-hotfix==0.6
Pygments==2.18.0
pyparsing==3.2.0
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tim==0.5
timm==1.0.12
tokenizers==0.15.2
tomli==2.2.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.4
traitlets==5.14.3
transformers==4.39.3
triton==3.0.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
urllib3 @ file:///croot/urllib3_1727769808118/work
wandb==0.17.7
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.0
